MSI: Move default stock manupulations to inventory catalog module
