+++ app/code/Magento/InventoryCatalog/Observer/SaveInventoryDataObserver.php
@@ -3,8 +3,11 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Inventory\Observer;
+namespace Magento\InventoryCatalog\Observer;
 
+use Magento\Catalog\Api\Data\ProductInterface;
+use Magento\Catalog\Api\ProductRepositoryInterface;
+use Magento\CatalogInventory\Model\Stock\Item;
 use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
 use Magento\InventoryApi\Api\Data\SourceItemInterfaceFactory;
@@ -12,6 +15,7 @@ use Magento\InventoryApi\Api\SourceItemsSaveInterface;
 use Magento\InventoryCatalog\Api\DefaultSourceProviderInterface;
 use Magento\Framework\Exception\CouldNotSaveException;
 use Magento\Framework\Exception\InputException;
+use Magento\Framework\Exception\NoSuchEntityException;
 
 /**
  * Synchronize quantity information which based on the CatalogInventory qty with
@@ -35,18 +39,26 @@ class SaveInventoryDataObserver implements ObserverInterface
     private $defaultSourceProvider;
 
     /**
+     * @var ProductRepositoryInterface
+     */
+    private $productRepository;
+
+    /**
      * @param SourceItemInterfaceFactory $sourceItemFactory
      * @param SourceItemsSaveInterface $sourceItemsSave
      * @param DefaultSourceProviderInterface $defaultSourceProvider
+     * @param ProductRepositoryInterface $productRepository
      */
     public function __construct(
         SourceItemInterfaceFactory $sourceItemFactory,
         SourceItemsSaveInterface $sourceItemsSave,
-        DefaultSourceProviderInterface $defaultSourceProvider
+        DefaultSourceProviderInterface $defaultSourceProvider,
+        ProductRepositoryInterface $productRepository
     ) {
         $this->sourceItemFactory = $sourceItemFactory;
         $this->sourceItemsSave = $sourceItemsSave;
         $this->defaultSourceProvider = $defaultSourceProvider;
+        $this->productRepository = $productRepository;
     }
 
     /**
@@ -54,14 +66,21 @@ class SaveInventoryDataObserver implements ObserverInterface
      * @return void
      * @throws InputException (thrown by SourceItemsSaveInterface)
      * @throws CouldNotSaveException (thrown by SourceItemsSaveInterface)
+     * @throws NoSuchEntityException (thrown by ProductRepositoryInterface)
      */
     public function execute(Observer $observer)
     {
-        /** @var \Magento\Catalog\Model\Product $product */
-        $product = $observer->getEvent()->getProduct();
-        $extendedAttributes = $product->getExtensionAttributes();
-        $stockItem = $extendedAttributes->getStockItem();
+        /** @var Item $item */
+        $item = $observer->getEvent()->getData('item');
 
+        /** @var ProductInterface $product */
+        $product = $this->productRepository->getById($item->getProductId());
+        $extendedAttributes = $product->getExtensionAttributes();
+        if (!$extendedAttributes) {
+            return;
+        }
+
+        $stockItem = $extendedAttributes->getStockItem();
         if (!$stockItem) {
             return;
         }
